Split functions into diff library and add md5 class function

+++ crawler.cpp
@@ -1,3 +1,7 @@
+// ADD fetch time, success rate, fail rate
+
+// later IP
+
 #include <string>
 #include <cstring>
 #include <sstream>
@@ -8,6 +12,10 @@
 #include <curlpp/Easy.hpp>
 #include <curlpp/Options.hpp>
 #include <curlpp/Exception.hpp>
+
+#include "md5.h"
+#include "rmTag.h"
+#include "url.h"
 
 #define MAX_FILE_LENGTH 50000
 
@@ -46,65 +54,66 @@ void write()
     output_file << "Content : " << std::endl << m_pBuffer << std::endl;
 }
 
-/* remove from <tag> to </tag>*/
-void rmTag(char *tag)
+/*void retrieveUrl()
 {
-  /* replace newline with  space */
-  for (int i=0; i<strlen(m_pBuffer); i++)
-    if(m_pBuffer[i]=='\n')
-      m_pBuffer[i] = ' ';
-  
-  char front_tag[20], back_tag[20];
-  sprintf(front_tag, "<%s", tag);   // create <tag for strstr
-  sprintf(back_tag, "</%s>", tag);  // create </tag> for strstr
-  char *front_ptr = m_pBuffer, *back_ptr=NULL;
-  int time=0;
-  while((front_ptr = strstr(front_ptr, front_tag))!=NULL)
-  {
-    std::cout << time++ << std::endl;
-    char *tmp = front_ptr;
-    front_ptr++;
-    *tmp = '\0';
-    if ((back_ptr = strstr(front_ptr, back_tag))==NULL) {
-      //error
-      std::cout << "error" << std::endl;
-      break;
-    }
-    std::cout << time++ << std::endl;
-    strcat(tmp, back_ptr+strlen(back_tag));
-    front_ptr--;
-  }
-  // std::cout << m_pBuffer << std::endl;
-}
-
-void retrieveUrl()
-{
-  char *href = m_pBuffer, url_link[128], *tmp;
+  char *href = m_pBuffer, url_link[256], *tmp;
   std::ofstream url_file;
   url_file.open("url.txt");
-  while((href=strstr(href, "href=\"http"))!=NULL)
+  while(1)
   {
+    int len = 0;
+    if ((href=strstr(href, "href=\"http"))==NULL)
+      break;
+    //std::cout << href << std::endl << std::endl;
     tmp = href = href+strlen("href=\"");
     while(1)
     {
-      if (*tmp=='\"') {
+      if (*tmp=='\"' || *tmp=='\0') {
         *tmp = '\0';
         break;
       }
       tmp++;
+      len++;
     }
-    strcpy(url_link, href);
+    memcpy(url_link, href, len+1);
+    //strcpy(url_link, href);
     *tmp = ' ';
     //std::cout << url_link << std::endl;
     url_file << url_link << std::endl;
   }
   url_file.close();
+}*/
+
+void retrieveBody()
+{
+  char body[4096];
+  char *head = strstr(m_pBuffer, "<body");
+  char *tail = strstr(m_pBuffer, "</body>"), *tmp = head+strlen("<body");
+  int len = 0;
+  while(1) {
+    if (*tmp=='>')
+      break;
+    tmp++;
+  }
+  head = tmp++;
+  while(1) {
+    len++;
+    if (strncmp(tmp, tail, 7)==0)
+      break;
+    tmp++;
+  }
+  *tail = '\0';
+  std::cout << head << std::endl;
+  //memcpy(body, head, len+10);
+  //strcpy(body, head);
+  *tail = ' ';
+  // std::cout << body << std::endl;
 }
 
 int main(int, char **) /* I/O for save data, using dataa batch to control I/O count */
 {
     m_pBuffer = (char*) malloc(MAX_FILE_LENGTH * sizeof(char));
-    //output_file.open("test.txt");
+    output_file.open("test.txt");
     try
     {
         // initial process
@@ -112,7 +121,7 @@ int main(int, char **) /* I/O for save data, using dataa batch to control I/O co
         curlpp::Easy myRequest;
 
         // create curl options
-        curlpp::options::Url myUrl(std::string("https://www.google.com"));
+        curlpp::options::Url myUrl(std::string("https://star.ettoday.net/news/718757"));
         curlpp::types::WriteFunctionFunctor functor(WriteMemoryCallback);
         curlpp::options::WriteFunction *write_func = new curlpp::options::WriteFunction(functor);
 
@@ -132,15 +141,19 @@ int main(int, char **) /* I/O for save data, using dataa batch to control I/O co
 	{
 		std::cout << e.what() << std::endl;
 	}
-    std::cout << "rm script" << std::endl;
-    rmTag("script");
-    std::cout << "rm style" << std::endl;
-    rmTag("style");
-    std::cout << "retrieve URL" << std::endl;
-    retrieveUrl();
-    std::cout << "write()" << std::endl;
-    //write();
-    //output_file.close();
-    free(m_pBuffer);
-    return 0;
+  if (m_Size==0)
+    return -1;
+  std::cout << "rm script" << std::endl;
+  rmTag(m_pBuffer , "script");
+  std::cout << "rm style" << std::endl;
+  rmTag(m_pBuffer, "style");
+  std::cout << "retrieve URL" << std::endl;
+  retrieveUrl(m_pBuffer);
+  std::cout << "retrieve body" << std::endl;
+  // retrieveBody();
+  std::cout << "write()" << std::endl;
+  write();
+  output_file.close();
+  free(m_pBuffer);
+  return 0;
 }
